+++ app/src/main/kotlin/akihz/anlaki/dev/AkihzApplication.kt
@@ -3,6 +3,9 @@ package akihz.anlaki.dev
 import android.app.Application
 import akihz.anlaki.dev.data.CustomProfileManager
 import akihz.anlaki.dev.data.UpdateDownloadStore
+import akihz.anlaki.dev.data.UpdateCheckScheduler
+import akihz.anlaki.dev.utils.PreferencesHelper
+import akihz.anlaki.dev.utils.UpdateAvailableNotification
 import akihz.anlaki.dev.utils.UpdateNotification
 import dagger.hilt.android.HiltAndroidApp
 import timber.log.Timber
@@ -11,7 +14,13 @@ import timber.log.Timber
 class AkihzApplication : Application() {
     override fun onCreate() {
         super.onCreate()
+        PreferencesHelper.init(this)
         CustomProfileManager.init(this)
+        if (PreferencesHelper.latestAvailableVersionCode <= BuildConfig.VERSION_CODE) {
+            PreferencesHelper.clearAvailableUpdate()
+            UpdateAvailableNotification.cancel(this)
+        }
+        UpdateCheckScheduler.sync(this)
         UpdateDownloadStore(this).load()
             ?.takeIf { it.update.versionCode <= BuildConfig.VERSION_CODE }
             ?.let {